Extra casts for extra silly compilers

+++ mixxx/src/vinylcontrolxwax.cpp
@@ -106,8 +106,8 @@ void VinylControlXwax::AnalyseSamples(short *samples, size_t size)
         timecoder_submit(&timecoder, samples, size);
 
         //Update the input signal strength
-        timecodeInputL->slotSet((float)fabs(samples[0]) / SHRT_MAX);
-        timecodeInputR->slotSet((float)fabs(samples[1]) / SHRT_MAX);
+        timecodeInputL->slotSet((float)fabs((float)samples[0]) / SHRT_MAX);
+        timecodeInputR->slotSet((float)fabs((float)samples[1]) / SHRT_MAX);
 
         waitForNextInput.wakeAll();
         lockSamples.unlock();
